requirements.txt
flask==3.1.2
geopy==2.4.1
google-cloud-logging==3.13.0
google-cloud-storage==3.7.0
numpy==2.4.0
pandas==2.3.3
shapely==2.1.2
unidecode==1.4.0
cryptography==43.0.1
